Input list patch (#252)
* Use the class we're testing

Not the deprecated version

* Re-insert Sam's recursive insertion

I think it just got lost in all the InputList madness. Strange though, since his tests made it through. Anyhow, I tried setting him as commit author, but this is the first time I've used that functionality in pycharm so forgive me if it doesn't work. -Liam

* Update comparison type in Sam's recursion test

* Make sure subclasses don't revert to DataContainer

Co-authored-by: Sam Waseda <37879103+samwaseda@users.noreply.github.com>