Pass SQL specific arguments
These are used by DatabaseAccess only and so should be removed.
I add them to guarantee uniform access right now, though.

sql_query should be demoted to a property of DatabaseAccess, so that it
no longer needs to be passed.

user is a bit more tricky, it seems it's only used by DatabaseAccess to
filter jobs, but it could be added to FileTable, too.  Then again there
should only be one user in a FileTable setup...