Simplify HDF5Stub
Instead of subclassing for each new lazily loadable type, types may
register themselves to the HDF5Stub class with a simple callback.  The
class then checks the 'NAME' field in the HDF group at loading time
against values provided in the register call and use this callback.

This allows for the same amount of customization, but has the advantage
that you can wrap every HDF5 group in HDF5Stub without checking which
subclass is necessary.