Save pyiron tables via DataContainer instead of csv
All data in the table are now saved via datacontainer and when loaded
from HDF5 the data is taken from there instead of from the csv file.
This makes sure complex data are correctly instantiated when reloading a
table, since writing to csv forces a text representation, which cannot
be converted back on loading.

This tests the new behavior for numpy
arrays, but structures, etc. are probably also affected.