Remove unnecessary conversion
This was silly. In the debug log the path being searched for is already in the right format, so this conversion is being safely done earlier. The problem is that the underlying config variable has the wrong format. It has the right format in the config file and at instantiation time, so it must somehow be getting overwritten later or running on a separate python instance where it's not getting read correctly to begin with.