It worked but...
I don't understand why. Converting the cwd call should absolutely have been necessary for that one test in TestManager, but I am concerned that it had spillover to any other tests. On my local machine we modify 'project_paths' but then restore it to its previous state without trouble. Is it possible that the github tests are being parallelized and whichever tests happen to be relying on settings at the time get broken somehow? I still find it very strange and somewhat disconcerting, but I'm grateful that it works and at least for the one test where I added the conversion it makes total sense why we would go from a fail to a pass...