Grab watchdog lock before checking connection status
Before we first check whether the connection is alive, then kick the
watchdog and only then grab the lock to make the SQL query.  There's a
race condition in this as sketched below, where the watchdog acquires
the lock and closes the connection after we check that it's still alive,
but before we acquire the lock.

```
Main Thread                     Watchdog Thread
-----------                     ---------------

                                TIMER RUNS OUT
                                        |
                                        v
CHECK CONNECTION        ~       ACQUIRE LOCK
    |                                   |
    |                                   v
    |                           CLOSE CONNECTION
    |                                   |
    |                                   v
    |                           RELEASE LOCK
    |
    v
ACQUIRE LOCK
    |
    v
EXECUTE SQL QUERY
    |
    v
RELEASE LOCK
```

The solution is to protect also the checking of the connection with the
same lock, so that closing of the connection happens only before or
after we checked the connection *and* executed our query.