Always write error.out
Previously error.out was written when the job was successful and
error.msg when not.  In case the job returned an accepted non zero
return nothing was written due to a bug.  This commit writes all the
output of the job to error.out in all cases, plus error.msg when the job
errors out for backwards compatibility.