Fix off-by-one in FlattenedStorage
When resizing a storage the number of allocated chunks are doubled, but
this breaks when there's no chunks pre-allocated (because, naja,
2*0==0...).  Normally this does not happen, because new storages start
with a pre-allocation of one, but in the case of an empty storage read
from HDF it does occur.  This now takes care to allocate at least one
chunk everytime.