Do not initialize Executable when none saved in HDF
The previous code first try'd to load saved executable information via

self.executable.from_hdf()

and then bailed if there was an error.  The problem is that just
accessing `self.executable` instantiates an `Executable` object on the
job even if it is a job, that does not need it, e.g. because it does not
require an external executable.  Due to the error this left the object
in an inconsistent state, unable to save itself again with `to_hdf()`.