Uniform database access to get child ids
Always use the get_child_ids() function on the project level, as this one already implements interfaces to both, the SQL database as well as the file based database. In addition for the file based database filter based on the working directory rather than the master_id, as no master_id exists in the file based table.