Allow to issue warning; special case __setattr__
Sub classes that use sentinel on __setattr__ need to special case when
the attribute to set is `read_only` or `_read_only`.  Failure to do so
would result in being unable to unlock the object.  `sentinel` now
handles this.

Additionally adds a `lock_method`.  Default is `error`, but can be set
to emit a warning and allow modification instead.