If JOB_CLASS_DICT contains class return early
Usually JOB_CLASS_DICT contains the module path from where to import the class.
However Project.create.job/register also supports live classes in there, but this was not
handled when loading jobs from HDF again.  That code path always expected the module path
(as a string) and consequently crashed when it found an object there.