Try to access files after HDF
Currently job[x] first checks whether x is a file in the directory of
job.  Crucially it also does this when the files are compressed and when
job was obtained in inspect mode.  Together this conspires to make
pyiron tables much slower, because every single table entry will
search the compress tar archive for (non-existing) files first.

This commit reorders the access so that first the HDF files is searched
for x, before looking for local files, based on two assumptions:
1. accessing HDF is the common case and should be faster
2. users that want to read files and avoid searching HDF can simply use
   job.files