Move recursive loading of DataContainer to HDF5Content
This will allow users and internal code to use job.content[...] if they what they are looking for
is saved in HDF5 storage and obviate the need to rely on job[...] which may be slow because it is
also searching the compressed job files.

I will need some logic like this to fix the slow loading of pyiron_atomistics jobs.