Introduce pyiron_mini
For users who just want to use pyiron as a workflow manager for any kind of python function or executable, `pyiron_mini` provides a minimal set of functions. You can import it as:
```python
import pyiron_base.mini as pyiron_mini
```
Then you have access to:
* `pyiron_mini.wrap_exe()` to wrap external executables.
* `pyiron_mini.wrap_py()` to wrap python functions.
* `pyiron_mini.cache_list()` to list the currently cached tasks.
* `pyiron_mini.cache_clear()` to clear the cache.
* `pyiron_mini.conda_create_env()` to create new conda environments to execute tasks in.

If `pyiron_mini` is not short enough you can also import it as `import pyiron_base.mini as pm` in that case it is as short as the `pr` we commonly use to reference a `Project` object.