Update integration tests and a new one for notebooks
Previously we tried to install the deps of atomistics/contrib on top of
the base environment, now we install those dependencies first and then
add pyiron_base HEAD without dependencies.
The former breaks when there are incompatible versions, the latter can
break if base actually needs features or whole modules from its own
environment, but we're betting that that's rarer.