Remove automatic recursive to_dict/from_dict again
This turned out to have negative interactions when writing and reading objects from GenericJob
which currently employs a mixture of to_hdf and newer to_dict interfaces.  When reading the
objects again the recursive strategy can get confused which interface to use.  Since it is
anyway mostly relevant for the DataContainer, I have extracted this functionality into stand-
alone functions that operate on the obj_dicts.  Classes that wish to use it and do not confuse
the two interfaces can call these functions in their implementations of _to_dict and _from_dict
as done by DataContainer now.