Only escape strings in SQL queries when % is issued by us
Only then will we actually create a LIKE statement and need to escape
any _ or % potentially present in user input.
By escaping all values always, I accidentally introduced a bug for
projects with _ in them, because then SQL sees the escaped project path
in the queries and takes them to be different from unescaped project
path in the database.